src/Python/fastapi_openai/pyproject.toml
[project]
name = "fastapi-openai"
version = "0.1.0"
description = "Add your description here"
authors = [
    { name = "aren", email = "code@arencore.me" }
]
dependencies = [
    "fastapi>=0.111.0",
    "black>=24.4.2",
    "uvicorn>=0.30.1",
    "sqlalchemy>=2.0.31",
    "asyncpg>=0.29.0",
    "psycopg2-binary>=2.9.9",
    "jinja2>=3.1.4",
    "openai>=1.35.13",
    "python-dotenv>=1.0.1",
]
readme = "README.md"
requires-python = ">= 3.8"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.rye]
managed = true
dev-dependencies = []

[tool.hatch.metadata]
allow-direct-references = true

[tool.hatch.build.targets.wheel]
packages = ["src/fastapi_openai"]
